fix: Bug where ticket value can be null
OnChannelDeleted event was expecting not null ticket when the channel topic was successful.

However this is only problem when bot can not create ticket but channel was created. When this bug occurs deleting channel was throwing exception.

Updated code to expect nullable ticket value and only act if it is not null.